+++ src/Vt102Emulation.h
@@ -100,7 +100,7 @@ protected:
     // reimplemented from Emulation
     void setMode(int mode) Q_DECL_OVERRIDE;
     void resetMode(int mode) Q_DECL_OVERRIDE;
-    void receiveChar(uint cc) Q_DECL_OVERRIDE;
+    void receiveChar(int cc) Q_DECL_OVERRIDE;
 
 private Q_SLOTS:
     // Causes sessionAttributeChanged() to be emitted for each (int,QString)
@@ -109,7 +109,7 @@ private Q_SLOTS:
     void updateSessionAttributes();
 
 private:
-    unsigned int applyCharset(uint c);
+    unsigned short applyCharset(unsigned short c);
     void setCharset(int n, int cs);
     void useCharset(int n);
     void setAndUseCharset(int n, int cs);
